requirements.txt
anthropic
apify-client
openpyxl
python-docx
reportlab
requests
python-dotenv
python-dateutil
beautifulsoup4
crawl4ai
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
